Last organiser cant be removed

+++ src/app/components/kandoe/themedetail/components/themedetail-organiser/themedetail-organiser.component.ts
@@ -66,15 +66,17 @@ export class ThemedetailOrganiserComponent implements OnInit {
   }
 
   removeOrganiser(user) {
-    this.themeService.removeUserFromTheme(user.id, this.themeId, this.userId).subscribe(
-      data => {
-        this.users = data;
-      },
-      error => {
-        console.error('Error removing user!');
-        console.log(error);
-        alert('Error removing user');
-      });
+    if (this.users.length > 1) {
+      this.themeService.removeUserFromTheme(user.id, this.themeId, this.userId).subscribe(
+        data => {
+          this.users = data;
+        },
+        error => {
+          console.error('Error removing user!');
+          console.log(error);
+          alert('Error removing user');
+        });
+    }
   }
 
 }
